move types dir to root
