+++ src/fetchers/file.js
@@ -1,5 +1,5 @@
 // @flow
-import type { ZelConfig } from '../types';
+import type { ZelConfig } from '../../types';
 const BaseFetcher = require('./base');
 
 class FileFetcher extends BaseFetcher {
